Updated Interactive
Not sure what the difference is between the interactive file and the index file, but with this version of the index file, once snapped into the response boxes, the cards can't move back to the original spots. In the interactive file, the cards can move back to the original slots after already snapped into the response slots.