Multiple Pages + Check Answer Button
When all the cards are placed correctly and the 'Check Answer' button is pressed, all of the incorrect cards are booted to their origin and if all of the cards are correct, a congratulatory message pops up. I also added multiple pages by splitting up the relevant sections of the original index.html file into round1/2 (commenting out the other rounds/filler paragraphs because I don't want to delete anything just yet). When the congratulatory message pops up in Round 2, the link to the next round appears. There are some minor formatting issues with the positions of the buttons/answer spoiler. I also started working on a 'Show Answer' button (which I currently labeled 'test' in the JavaScript function) but there seems to be something wrong with how the positions are stored because the cards move to different positions (when the function currently tries to move them all to the same place) and in DevTools, the position given for the cards in 'Elements' varies greatly with the position they actually move to, shown in 'Console', as per the function which logs the location after animation.