(hopefully) Final Edits
The 'show answer' button now works! Cards move to their correct spaces when the button is clicked. If a correct card is already in place, the other draggable cards will move to the empty spots. If a card is in the wrong spot, it will move to the correct spot when the button is clicked. The 'restart' button refreshes everything to make the game playable again.

Aside from the button tweaks, I also finalized the formatting and styling so everything looks neat and consistent.

Certain elements (answer spoiler, and link to continue) pop up when the 'check answer' and/or 'show answer' buttons are clicked, and the formatting is neat, but could potentially be less cluttered (not a huge issue though).