Fe's Picture + Copyright/Citation + CSS Style Sheet for Spanish Pages
I've added the picture of Fe + relevant text for the English and Spanish intro pages. I've also added copyright/citation notes at the end of round2 for the English version (don't want to butcher the Spanish translation). I've also added the GoT CSS style sheet for the Spanish version of the game